Change bg texture. WIP symposium layout changes.
FIX. Change the background texture to look like a rock more closely. The
background texture was made with inkscape and edited by hand to make use
of the proper filter.

WIP. Changes to the layout to accomodate smaller screens. Right now, the
flex layout works for chrome on larger screens but goes completely crazy
on firefox and safari on smaller screens.